+++ uui/components/datePickers/DatePicker.tsx
@@ -10,7 +10,12 @@ import { DropdownContainer } from '../overlays';
 const defaultMode = EditMode.FORM;
 
 /** Represents the properties of the DatePicker component. */
-export interface DatePickerProps extends DatePickerCoreProps, SizeMod, IHasEditMode {}
+export interface DatePickerProps extends DatePickerCoreProps, SizeMod, IHasEditMode {
+    /**
+    * HTML ID attribute for the toggler input
+    */
+    id?: string;
+}
 
 class DatePickerComponent extends BaseDatePicker<DatePickerProps> {
     renderInput = (props: IDropdownToggler & { cx: any }) => {
@@ -46,6 +51,7 @@ class DatePickerComponent extends BaseDatePicker<DatePickerProps> {
                 onBlur={ this.handleBlur }
                 mode={ this.props.mode || defaultMode }
                 rawProps={ this.props.rawProps?.input }
+                id={ this.props.id }
             />
         );
     };